requirements.txt
requests>=2.31.0
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
feedparser>=6.0.0
rapidfuzz>=3.0.0
rich>=13.0.0
schedule>=1.2.0
python-telegram-bot>=20.0
